/branches/1.3.x:
Merged revision 6131 from trunk.

r6131 | sa | 2009-09-30 23:15:54 CEST

trunk:

If a merge has a delete followed by an add on the same 
node, and also that node is modified locally, svn tried 
to create two tree-conflicts on the same node. That hit 
an assertion and bailed out, breaking the working copy.

SVNMergeDriver:
 - added recordTreeConflictOnAdd()
 
SVNMergeCallback:
 - fileAdded(), directoryAdded(): call 
   myMergeDriver.recordTreeConflictOnAdd() instead of 
   myMergeDriver.recordTreeConflict() where necessary.
   
SVNRemoteDiffEditor:
 - SVNEvent notifications changed for addDir(), closeFile() 
   to meet requirements of the fixed use-case.
----------------------------------------------------------------------------